server/http/requirements.txt
aio-pika==1.7.1
aiofiles==0.3.1
argon2-cffi==18.1.0
asyncpg==0.11.0
cffi==1.11.2
decorator==4.0.11
httptools==0.0.9
jsonpath-rw==1.4.0
jsonschema==2.6.0
multidict==3.3.2
passlib==1.7.1
pid==2.1.1
pika==0.10.0
ply==3.10
pycparser==2.18
PyJWT==1.5.3
sanic==0.7.0
sh==1.12.14
shortuuid==0.5.0
six==1.11.0
ujson==1.35
uvloop==0.9.1
websockets==3.3
yarl==0.16.0
